terror: refactor the file struct
